Added PID controller for the watermodule
works. Not perfect but it is functional to keep a setpoint.

+++ WaterModule/TempSensor.cpp
@@ -1,7 +1,7 @@
 #include "TempSensor.h"
 #include "Arduino.h"
 
-int TempSensor::GetData()
+double TempSensor::GetData()
 {
   double sensorvalue = analogRead(SensorPin);
